Remove usage of JDOM library
Use Commons Configuration to read/write the XML files instead of JDOM,
the former is used in more places.
The library will still be made available to add-ons, until they are
changed to bundled it or use other library.
Add tests to assert that the corresponding XML files are correctly
handled.
Correct typos in XML elements of the drivers (no impact on functionality
the value of the element had the default value).

Part of #4780 - Tidy up dependencies

Signed-off-by: thc202 <thc202@gmail.com>