Add task to build the weekly release
Add task `distWeekly` that creates the weekly release, automatically
building the required add-ons (the task will clone, or update if already
cloned, the required repositories).
Add task `copyWeeklyAddOns` to build and copy the weekly add-ons from
existing repositories, in the file system at the same level as zaproxy.
The add-ons are copied to the (dist) plugin directory.

Signed-off-by: thc202 <thc202@gmail.com>