Initial CustomPages Implementation
Add support for custom error page definitions associated with a Context.
Expose those details and functionality to Passive and Active scanners
(rules). Assume that if a context with custom pages is setup the user
wants those details to be leveraged. Default behavior is to check custom
pages and fallback to status code checking (based on the applicable
custom page definition type and related status codes).

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>