Parse result of PAC script evaluation
Add method `findProxyForUrl` that calls the script function and parses
its result, returning a list with the proxy settings.
Reduce the visibility of `evaluate`, the entry point is now the added
method.
Add constructors for `Path` and `String`, needed when creating the
script from a file and from the contents of a HTTP response.

Part of #265 - Allow to configure outgoing proxy with PAC file/URL

Signed-off-by: thc202 <thc202@gmail.com>