Default to author/URL of the add-on in Extension
Change `Extension` to default to author/URL of the add-on when set.
Remove override `ExtensionAdaptor#getURL()`, no longer needed.
Remove unnecessary null check when setting the author to the text
component, setting null or empty string has the same effect.

Avoids the need to override those methods for most extensions provided
by add-ons, since the author and URL (when set) are usually the same.

Signed-off-by: thc202 <thc202@gmail.com>