Allow add-ons to declare/bundle its dependencies
Change `ZapAddOnXmlFile` to read the libraries from the manifest,
declared as:
```xml
<libs>
  <lib>libs/lib1.jar</lib>
  <lib>libs/lib2.jar</lib>
</libs>
```
Each `lib` entry has the path of the library in the add-on/JAR, which
will be copied to the file system under the following path:
`<zapHome>/addOnData/<addOnId>/libs/`
(The directory structure declared in the add-on is not maintained.)

Change `AddOn` to:
 - Read/provide the libs contained in the manifest, so that they can be
   used elsewhere (e.g. un/installed, added to classloader);
 - Validate that the libraries are present in the add-on;
 - Check that the libraries are already present in the file system when
   verifying if the add-on is runnable;
 - Provide a method to check if the add-on/extensions can be installed
   (same check as runnable without checking if the libraries are already
   in the file system).

Change `AddOnCollection` and `AddOnDependencyChecker` to call the new
method that checks if the add-on/extensions can be installed.
Change `ExtensionAutoUpdate` to do the same and report invalid library
issue.
Change `AddOnRunIssuesUtils` to provide the error message when the
libraries are missing in the add-on.
Change `AddOnInstaller` to allow to install/uninstall the libraries of
an add-on and install missing libraries (in case a library was removed
after installing the add-on).
Change `AddOnClassLoader` to allow to add a list of URLs, to also search
for classes/resources in the add-on libraries.
Change `AddOnLoader` to:
 - Install and uninstall the libraries when the add-on is un/installed;
 - Install the missing libraries of an already installed add-on before
 checking if it's runnable;
 - Set the libraries to the respective `AddOnClassLoader`.

Fix #2619.

Signed-off-by: thc202 <thc202@gmail.com>