Expose Context details to passive scan rules
PassiveScanData > A new helper class which exposes the utility methods
(getUsers(), and hasContext()) for the HttpMessage currently being
passively scanned.
PluginPassiveScanner > Added new member and getter (getHelper()) for the
PassiveScanData (which is set in the init() method).
PassiveScanThread > Instantiates a PassiveScanData variable with the
current HttpMessage and passes it to PluginPassiveScanner.init().

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>