Add tasks to generate JFlex TokenMakers
Add task to run JFlex (latest, v1.8.2) for given lexers.
Add task to change the JFlex class to RSyntaxTextArea's `TokenMaker`, to
avoid doing the required changes manually.

Configure the build to create tasks for lexers that match the name
`*TokenMaker.flex` under `src/main/flex`, which use the tasks mentioned
earlier.
For example, for `WwwFormTokenMaker.flex` it is created the task:
 - `:zap:generateRstaWwwFormTokenMaker`

All the classes can be generated with:
 - `:zap:generateRstaTokenMakers`

Include `TokenMaker` classes in Spotless for consistent format and
indentation (after generation and automatic changes).
Ignore `TokenMaker` classes when doing binary compatibility checks,
those classes are not intended to be used by add-ons.

Update `WwwFormTokenMaker` lexer:
 - Remove JavaDoc with description of changes, now automatically done;
 - Remove constructor and methods no longer needed;
 - Correct package name;
 - Tweak imports;
 - Address warnings and update `yyreset` for newer JFlex version;
 - Extend `AbstractJFlexTokenMaker`, better base class.

Regenerate the class and reformat.

Signed-off-by: thc202 <thc202@gmail.com>