Catch NPE on panel level when saving/validating
Move the catch of the NullPointerException from the container panel to
the individual panels to still allow to validate/save other panels that
do not have issues.
Make the error generic now used for save/validate and provide the
details of the exception for easier report.

Related to #6136.
Improvement of #5306.

Signed-off-by: thc202 <thc202@gmail.com>