Test common functions of packaged scans
Test the following common functions:
 - `zap_spider`
 - `zap_ajax_spider`
 - `zap_active_scan`
 - `zap_wait_for_passive_scan`
 - `zap_import_context`

Change a log statement in `zap_ajax_spider` to not concatenate the log
message to cope with mocked objects.
Change `zap_import_context` to check that the ID of the context was
returned (int) instead of a specific error message, not the same for all
possible errors.

Signed-off-by: thc202 <thc202@gmail.com>