Improve error message when starting packaged scans
Fail immediately if the spider scans were not started to provide better
error message, show:
`Failed to start the scan, check the log/output for more details.`

instead of, e.g.:
`invalid literal for int() with base 10: 'does_not_exist'`.

when checking the status of the scan.

Correct `suite.py` to exit with expected code when tests fail.

Signed-off-by: thc202 <thc202@gmail.com>