Address random test failures
Change `ExtensionSessionManagementUnitTest` to reset the state of the
`ScriptBasedSessionManagementMethodType` by setting the script extension
to null (none), it could be already initialised by other test causing
them to fail (trying to register the script type again).
Change `PausableScheduledThreadPoolExecutorUnitTest` to run the tasks
with more delay to give more time for the tasks to execute in case of
slowdowns.

Signed-off-by: thc202 <thc202@gmail.com>