Address some glitches when changing look and feel
Update global pop up menus, otherwise some of them could still be shown
with previous look and feel.
Change help button in param panels to have an empty border instead of
null, to prevent the look and feel change from resetting it.
Related to help button (but not UI glitch), use the same help icon
already loaded by the help extension (and in the extension itself).
Replace toolbar used as spacer which could be detached after changing
the look and feel.

Part of #6201.

Signed-off-by: thc202 <thc202@gmail.com>