Warn when unable to save malformed HTTP message
Change models to throw an exception if unable to create the HTTP headers
instead of just logging a warning and continue, to allow to show the
error to the user.
Change Manual Request Editor and Break to warn if the data was not
actually set into the message.
Warn when unable to switch components, views, and HTTP method.

Fix #4235.

Signed-off-by: thc202 <thc202@gmail.com>