Add node for URLs ending with a slash
This allows us to differentiate between https://www.example.com/a and
https://www.example.com/a/

The main impact is to the active scanner as this works off the sites
tree - if you select the new "/" node then it will have no children and
therefore not recurse as most people will want.
So now if you specify a URL wit ha trailing slash then the slash will be
ignored if you specify 'recurse'.

Will require a corresponding change to the Quick Start add-on

Signed-off-by: Simon Bennetts <psiinon@gmail.com>