Do not require scan rules to implement setParent
Add implementation in `PluginPassiveScanner` to not require the scan
rules to implement it (most no longer need it), moreover allow to add
tags to the messages with a method in the scan rule itself to not
require the parent (directly) anymore.
Add method to `PassiveScanThread` to add a tag to a message and
deprecate existing one which has an unused parameter.
Change core passive scanners to no longer use the deprecated method.
Allow to extend `PassiveScanData` to ease the tests.

Signed-off-by: thc202 <thc202@gmail.com>