Add option for allowing app integration in containers
And use the option for launching the system browser.
If ZAP appears to be running in a container then openning a URL in the
system browser will be disabled - a tool tip will explain why.
The new display option can be used to re-enable it - the user could have
made some changes which mean this will actually work.

Signed-off-by: Simon Bennetts <psiinon@gmail.com>