Fix unit test for ApiUtils.getIntParam exceptions
Many unit tests in the `ApiUtilsUnitTest` class follow the following
pattern:
```
try {
    // Given
    // Set up some faulty params

    // When
    ApiUtils.getXyzParam(params, name);
} catch (ApiException e) {
    // Then
    assertThat(e.getType(), some_logic)));
}
```

If `getXyzParam` throws an exception, this pattern properly asserts
its details, but if the method fails to throw an exception at all
(e.g., due to a bug), no asserting will be called and the test will
pass.
E.g., if `getIntParam` is replaced with a single line of
`return -1;`, all these tests will pass, although they clearly
shouldn't.

A stronger pattern would be to use JUnit Jupiter's `assertThrows`, as
the `shouldThrowNullPointerExceptionWhenGettingBooleanFromNullParams`
does.

This patch fixes the other tests and aligns them to a similar
implementation as
`shouldThrowNullPointerExceptionWhenGettingBooleanFromNullParams`.

Signed-off-by: Allon Mureinik <allon@synopsys.com>