Correct encoding/decoding in URL path variant
Do not rely on `URLEncoder` to encode the path segments as that's not
the correct encoder for the URL path.
Handle and decode the path segments separately to preserve the encoded
values on each segment.
Keep the segments extracted when parsing the URL path to avoid doing it
each time a parameter is set.
Remove commented code.

Fix #5935.

Signed-off-by: thc202 <thc202@gmail.com>