Add support for mandatory add-ons
Some add-ons might be needed for ZAP to be actually functional, those
add-ons will be defined in the code and automatically set as mandatory.
When not found ZAP will fail to start, exiting with appropriate error
message.
The extensions of mandatory add-ons will be always enabled.
The mandatory add-ons can't be uninstalled, only updated.

Signed-off-by: thc202 <thc202@gmail.com>