Use latest proxy settings always
Change `HttpSender` to always use the latest proxy settings, also, use
a single client instance.
Previously it would keep using the old settings from when the
`HttpSender` was instantiated, e.g. the user would have to close the
manual request editor and open it again for the proxy settings to take
effect.

Signed-off-by: thc202 <thc202@gmail.com>