Restore Session title and Persist Session's state
Properly initialise the view in `ExtensionUiUtils` to have the session
changed listener added on hook, which sets the main title to the
session's name and controls the state of Persist Session menu/button.
Remove statements no longer needed in `MainMenuBar` that updated the
button state, now done in a single method.

Fix #7423.

Signed-off-by: thc202 <thc202@gmail.com>