Remove add-on `ResourceBundle` on soft uninstall
Remove the `ResourceBundle` when soft uninstalling, it's added back
when reinstalled, same logic as when uninstalling the add-on except for
the add-on files which are not removed.
This could cause log of errors during the update of the add-ons (should
not affect functionality, it would use the same data), e.g.:
`ERROR ….I18N - Adding message bundle with duplicate prefix: retire`

Part of zaproxy/zaproxy#7765.

Signed-off-by: thc202 <thc202@gmail.com>