Ensure passive scan runs even with high traffic
Obtain the ID of the last message to scan after any interruption.

The `PassiveScanController` is interrupted when responses are received,
with high/constant traffic it can be interrupted often that it would
not be able to get the ID of the last message after the first interrupt
preventing the new messages from being scanned.
This only happens with new sessions, ones that had no messages
persisted.

Signed-off-by: thc202 <thc202@gmail.com>