Fix release of macOS dist in CI
Move the creation of the symlink to the DMG task, being in the copy
task makes Gradle check it, leading to out of memory errors.
Correct update of the release:
 - Search the release by name if not found by tag, since it might still
 be in draft state.
 - Use expected newline char in the table header, otherwise the
 checksums would not be included.

Increase the max memory in the release tasks to ensure there are no
memory errors when creating and uploading the dists/installers.

Signed-off-by: thc202 <thc202@gmail.com>