Skip deleted msgs in the messages API endpoints
Handle messages that no longer exist when processing the messages for
one of the messages API endpoints.
It's possible that from the moment the IDs of the messages were
obtained to when the messages are actually read from the DB that they
no longer exist (e.g. spider tasks or other temporary messages were
deleted in the meantime).

Signed-off-by: thc202 <thc202@gmail.com>