Allow moving tabs to new windows
Add a context menu item to the tabs in all tabbed panels
(`TabbedPanel2`) that moves the tab to a new window.

Closing the window puts the tab back in its original position.

The list of detached tabs is not currently persisted and tabs must be
detached again when ZAP is restarted.

Signed-off-by: ricekot <github@ricekot.com>
Signed-off-by: ricekot <git@ricekot.com>