Fix Docker httpsender script
Fixes an issue in the Alert_on_HTTP_Response_Code_Errors.js included in
the zap-api-scan.py script.

When .setEvidence is called to record the status code returned by the
target server the return value from a regexp match was used. This
returns an array of strings which raises an error as the function
expects a string.

The status code is already extracted from the message earlier in the
script, so that value is used instead of parsing it from the header.

This resolves #9139.

Signed-off-by: Johan Svan <johan@svan.dev>